P0: preserve translation save error flow
